+++ packages/pyroscope-flamegraph/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.14.3](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.14.2...@pyroscope/flamegraph@0.14.3) (2022-07-15)
+
+**Note:** Version bump only for package @pyroscope/flamegraph
+
+
+
+
+
+## [0.14.2](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.14.1...@pyroscope/flamegraph@0.14.2) (2022-07-15)
+
+**Note:** Version bump only for package @pyroscope/flamegraph
+
+
+
+
+
+## [0.14.1](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.14.0...@pyroscope/flamegraph@0.14.1) (2022-07-15)
+
+**Note:** Version bump only for package @pyroscope/flamegraph
+
+
+
+
+
+# [0.14.0](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.13.0...@pyroscope/flamegraph@0.14.0) (2022-07-15)
+
+
+### Features
+
+* update right-click context menu ([#1259](https://github.com/pyroscope-io/pyroscope/issues/1259)) ([8aea02f](https://github.com/pyroscope-io/pyroscope/commit/8aea02f56320daacfd753d73db6936dcc7cdaef8))
+
+
+
+
+
 # [0.13.0](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.12.5...@pyroscope/flamegraph@0.13.0) (2022-07-13)