+++ packages/pyroscope-panel-plugin/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.28](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.3.27...@pyroscope/panel-plugin@1.3.28) (2022-07-15)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
+## [1.3.27](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.3.26...@pyroscope/panel-plugin@1.3.27) (2022-07-15)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
+## [1.3.26](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.3.25...@pyroscope/panel-plugin@1.3.26) (2022-07-15)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
+## [1.3.25](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.3.24...@pyroscope/panel-plugin@1.3.25) (2022-07-15)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
+## [1.3.24](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.3.23...@pyroscope/panel-plugin@1.3.24) (2022-07-15)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
 ## [1.3.23](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.3.22...@pyroscope/panel-plugin@1.3.23) (2022-07-13)
 
 **Note:** Version bump only for package @pyroscope/panel-plugin