+++ webapp/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.21.3](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.21.2...@pyroscope/webapp@1.21.3) (2022-07-15)
+
+**Note:** Version bump only for package @pyroscope/webapp
+
+
+
+
+
+## [1.21.2](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.21.1...@pyroscope/webapp@1.21.2) (2022-07-15)
+
+**Note:** Version bump only for package @pyroscope/webapp
+
+
+
+
+
+## [1.21.1](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.21.0...@pyroscope/webapp@1.21.1) (2022-07-15)
+
+**Note:** Version bump only for package @pyroscope/webapp
+
+
+
+
+
+# [1.21.0](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.20.0...@pyroscope/webapp@1.21.0) (2022-07-15)
+
+
+### Features
+
+* update right-click context menu ([#1259](https://github.com/pyroscope-io/pyroscope/issues/1259)) ([8aea02f](https://github.com/pyroscope-io/pyroscope/commit/8aea02f56320daacfd753d73db6936dcc7cdaef8))
+
+
+
+
+
 # [1.20.0](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.19.3...@pyroscope/webapp@1.20.0) (2022-07-13)